scripts/pyproject.toml
[project]
name = "ahrism-graph-builder"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.5.0",
    "konlpy>=0.6.0",
    "torch>=2.0.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
